py-backend/app/utils/assistant/model_selector.py
import os
from ollama import AsyncClient as OllamaAsyncClient
from google import genai # Assuming you'd use a suitable Gemini library

# Configuration for Ollama/Gemini - load these from a config file or environment variables
# For simplicity, they are hardcoded here.
OLLAMA_LOCAL_HOST = "http://localhost:11434"
OLLAMA_CLOUD_HOST = os.environ.get("OLLAMA_CLOUD_HOST")
GEMINI_API_KEY = os.environ.get("GEMINI_API_KEY")

class LLMService:
    """A service class to manage different LLM clients."""
    def __init__(self):
        # Initialize clients when the app starts
        self.ollama_local_client = OllamaAsyncClient(host=OLLAMA_LOCAL_HOST)
        # Assuming a Cloud Ollama client is similar or an external API client
        # self.ollama_cloud_client = ...

        # Initialize the Gemini client
        # self.gemini_client = genai.Client(api_key=GEMINI_API_KEY) 

    def get_llm_client(self, model_identifier: str, is_cloud: bool):
        """Returns the appropriate LLM client based on model and configuration."""
        
        # 1. Ollama Models
        if model_identifier.startswith("ollama:"):
            if is_cloud and OLLAMA_CLOUD_HOST:
                # Placeholder for a cloud-based Ollama setup
                # For this example, we'll return the local one, but you'd swap the client
                return self.ollama_local_client # Replace with cloud client
            else:
                return self.ollama_local_client

        # 2. Gemini Models
        elif model_identifier.startswith("gemini:"):
            # if self.gemini_client:
            #     return self.gemini_client
            # raise ValueError("Gemini client not configured.")
            
            # Placeholder for Gemini logic
            return None # You would return your Gemini client here

        else:
            raise ValueError(f"Unknown model identifier or configuration: {model_identifier}")

llm_service = LLMService()

async def get_ollama_stream(client, model: str, messages: list):
    """Abstraction for Ollama streaming."""
    return await client.chat(
        model=model.split(":")[1], # e.g., get 'gemma2:9b' from 'ollama:gemma2:9b'
        messages=messages,
        stream=True
    )

async def get_gemini_stream(client, model: str, messages: list):
    """Placeholder for Gemini streaming logic."""
    # response = client.models.generate_content_stream(...)
    # return response
    pass