py-backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asyncpg==0.30.0
bcrypt==5.0.0
certifi==2025.11.12
click==8.3.1
ecdsa==0.19.1
fastapi==0.121.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ollama==0.6.1
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
